fix: Use only one webhook per channel
Discord limits channels to 15 webhooks. These are shared between everyone
in the channel, so creating lots of webhooks not only screws up ourselves,
but it also messes with other bots that use webhooks, like Pluralkit.
The correct way to approach this problem is to use only one webhook per
channel, and use the `username` and `avatar_url` parameters to distinguish
different llms.